Move in favor of hash of config object
Config felt overly complex for what is going on here. Might bring it back but I don't think we need it just yet.
