+++ lib/flipper/adapters/http.rb
@@ -1,75 +1,11 @@
 require 'net/http'
 require 'json'
 require 'set'
+require 'flipper'
+require 'flipper/adapters/http/request'
 
 module Flipper
   module Adapters
-    # class for handling http requests.
-    # Initialize with Configuration instance
-    # Configuration attributes will be sent in every request
-    class Request
-      DEFAULT_HEADERS = {
-        'Content-Type' => 'application/json',
-        'Accept' => 'application/json',
-      }.freeze
-
-      def initialize(configuration)
-        @headers = DEFAULT_HEADERS.merge(configuration.headers.to_h)
-        @basic_auth_username, @basic_auth_password = configuration.basic_auth.to_h.first
-        @read_timeout = configuration.read_timeout
-        @open_timeout = configuration.open_timeout
-      end
-
-      # Public: GET http request
-      def get(path)
-        uri = URI.parse(path)
-        http = net_http(uri)
-        request = Net::HTTP::Get.new(uri.request_uri, @headers)
-        request.basic_auth(@basic_auth_username, @basic_auth_password) if basic_auth?
-        http.request(request)
-      end
-
-      # Public: POST http request
-      def post(path, data)
-        uri = URI.parse(path)
-        http = net_http(uri)
-        request = Net::HTTP::Post.new(uri.request_uri, @headers)
-        request.body = data.to_json
-        request.basic_auth(@basic_auth_username, @basic_auth_password) if basic_auth?
-        http.request(request)
-      end
-
-      # Public: DELETE http request
-      def delete(path, data = {})
-        uri = URI.parse(path)
-        http = net_http(uri)
-        request = Net::HTTP::Delete.new(uri.request_uri, @headers)
-        request.body = data.to_h.to_json
-        request.basic_auth(@basic_auth_username, @basic_auth_password) if basic_auth?
-        http.request(request)
-      end
-
-      private
-
-      def basic_auth?
-        @basic_auth_username && @basic_auth_password
-      end
-
-      def net_http(uri)
-        http = Net::HTTP.new(uri.host, uri.port)
-        http.read_timeout = @read_timeout if @read_timeout
-        http.open_timeout = @open_timeout if @open_timeout
-        http
-      end
-    end
-
-    class Configuration
-      attr_accessor :headers, :basic_auth, :read_timeout, :open_timeout
-    end
-
-    # Flipper API HTTP Adapter
-    # uri = URI('http://www.app.com/mount-point')
-    # Flipper::Adapters::Http.new(uri)
     class Http
       include Flipper::Adapter
       attr_reader :name
@@ -83,22 +19,11 @@ module Flipper
         end
       end
 
-      class << self
-        attr_accessor :configuration
-      end
-
-      # Public: initialize with api url
-      # http://www.myapp.com/api-mount-point
-      def initialize(uri)
-        configuration = self.class.configuration || Configuration.new
-        @request = Request.new(configuration)
-        @uri = uri.to_s
+      def initialize(options = {})
+        @options = options
+        @request = Request.new(options)
+        @uri = options.fetch(:uri).to_s
         @name = :http
-      end
-
-      def self.configure
-        self.configuration ||= Configuration.new
-        yield(configuration)
       end
 
       def get(feature)
